requirements.txt
flask
pandas
torch
faiss-cpu
transformers
groq
